@@ -1,6 +1,6 @@
 module go.abhg.dev/gs
 
-go 1.23.4
+go 1.24.0
 
 require (
 	github.com/alecthomas/kong v1.8.1
